Fix the macro button
There's a lot more global macros now, so we weren't finding a
previously-created macro. Also, most of the spells we used no longer
exist, so get rid of all of the dead spells. I don't have any spells to
replace them with at the moment.

Because most of the spells are dead, also define macro rules for cases
where we only have an in-combat spell and not an out-of-combat one.